Bytt over til nytt topic i prod. Bytter navn i tråd med navnebytte fra Brukernotifikasjon -> MittNav varsler
