+++ src/main/resources/application-dev-gcp.properties
@@ -10,7 +10,4 @@ fptilbake.scopes=api://dev-fss.teamforeldrepenger.fptilbake/.default
 saf.base.url=https://saf-q2.dev-fss-pub.nais.io
 saf.scopes=api://dev-fss.teamdokumenthandtering.saf/.default
 
-minside.kafka.topic.oppgave=min-side.aapen-brukernotifikasjon-oppgave-v1
-minside.kafka.topic.beskjed=min-side.aapen-brukernotifikasjon-beskjed-v1
-minside.kafka.topic.done=min-side.aapen-brukernotifikasjon-done-v1
 foreldrepenger.innsynlenke=https://foreldrepenger.intern.dev.nav.no
