+++ src/main/resources/application-prod-gcp.properties
@@ -1,4 +1,1 @@
-minside.kafka.topic.oppgave=min-side.aapen-brukernotifikasjon-oppgave-v1
-minside.kafka.topic.beskjed=min-side.aapen-brukernotifikasjon-beskjed-v1
-minside.kafka.topic.done=min-side.aapen-brukernotifikasjon-done-v1
 foreldrepenger.innsynlenke=https://foreldrepenger.nav.no
