+++ src/main/resources/application-vtp.properties
@@ -33,6 +33,7 @@ fptilbake.scopes=api://vtp.teamforeldrepenger.vtp/.default
 fptilbake.base.url=http://fptilbake:8080/fptilbake
 
 kafka.topic.journal.hendelse=teamdokumenthandtering.aapen-dok-journalfoering-vtp
+kafka.topic.minside.brukervarsel=ikkebrukt
 
 NAIS_DATABASE_FPOVERSIKT_FPOVERSIKT_HOST=localhost
 NAIS_DATABASE_FPOVERSIKT_FPOVERSIKT_PORT=5432
@@ -53,7 +54,4 @@ KAFKA_CREDSTORE_PASSWORD=changeit
 
 #server.port=8889
 
-minside.kafka.topic.oppgave=ikkebrukt
-minside.kafka.topic.beskjed=ikkebrukt
-minside.kafka.topic.done=ikkebrukt
 foreldrepenger.innsynlenke=http://localhost:9066/
