+++ src/main/resources/application.properties
@@ -3,4 +3,4 @@ application.name=fpoversikt
 
 kafka.topic.journal.hendelse=teamdokumenthandtering.aapen-dok-journalfoering
 
-minside.kafka.varseltopic=min-side.aapen-brukervarsel-v1
+kafka.topic.minside.brukervarsel=min-side.aapen-brukervarsel-v1
